Do not require sentence to be moved into SentenceWithPieces
Before this change `SentenceWithPieces` always contained an owned
`Sentence`. With this change, `SentenceWithPieces` can also hold
a mutable reference to a `Sentence`.

In order to support this in the `Tokenize` trait, it now has
two method:

- `tokenize`: takes ownership of a `Sentence` and stores it in
  the returned `SentenceWithPieces`.
- `tokenize_mut`: takes a mutable reference to `Sentence` and
  stores it in `SentenceWithPieces`.

Ideally, we would always use references. Unfortunately, this
does not work for iterators with `SentenceWithPieces` items.