Use the AdamW optimizer provided by torch
The tch binding now has support for the AdamW opt and for parameter
groups. Consequently, we do not need our own AdamW optimizer
implementation anymore. Switching to the Torch optimizer also speeds
up training a bit.

In tch the group is set by number on a `Path`. This gives us two
options:

1. Update syntaxdot-transformers to set the groups. This is quite
   inelegant, since we would decide for the caller what groups
   are used.

2. Create a wrapper for `Path`, `PathExt`, which takes a function that
   maps a variable name to a group. This allows the called to decide
   what variables should be put in what groups.

I have decided to go for (2). This introduces the `tch-ext` crate in
which `PathExt` is implemented.