distill: add the `attention-loss` option
Enabling this option adds the mean squared error (MSE) of the teacher
and student attentions to the loss. This can speed up convergence,
because the student learns to attend to the same pieces as the teacher.

Attention loss can only be computed when the teacher and student have
the same sequence lengths. This means practically that they should use
the same piece tokenizers.