Rename BertLayerOutput to LayerOutput and make it an enum
BertLayerOutput could either represent the embeddings of a batch
or the output of an encoder layer of a batch. This distinction
was somewhat obscured and could only be found out by checking if
the layer had attention.

This change, besides renaming BertLayerOutput to LayerOutput,
changes the layer output type to an enum with two variants:

- Embedding: the output of an embeddinglayer
- EncoderWithAttention: the output of an encoder layer.

Some getters are provided for uniform access to e.g. the layer
output.