Remove support for loading pretrained models from HDF5
Instead, we load pretrained models from files in the libtorch
OutputArchive format. This has two major advantages:

- This removes the HDF5 dependency, plus a lot of code for handling
  models in HDF5 format.
- Loading of pretrained models now share code paths with all other
  model loading.

The conversion scripts are updated to output pretrained embeddings
in the correct format.

Warning: this change breaks backwards compatibility with existing
BERT-based SyntaxDot models (e.g. BERT and XLM-RoBERTa). The incorrect
prefix "encoder" was used for embeddings. This was a known bug, but
now is a good moment to rectify this.