Add basic support for token pooling
This change adds a data type PiecePooler that is used by BertModel to
pool the pieces of a token after encoding of the input. This has
several benefits:

- Every layer after the pooler can assume that their input consists of
  token representations. This simplifies those layers, because they do
  not have to mask out non-initial pieces.
- It opens up the possibility to implement and experiment with
  different pooling functions.

Currently, only one pooler is available: PiecePooler::Discard. This
pooler uses the initial piece of a token as the token representation
and discards all continuation pieces. It results in the same
representations being used prior to this change and is therefore
compatible with existing SyntaxDot 0.3 models