Add mean pooler for token representations
This pooler takes the mean of all the representations of the pieces in
a token as the token's representation.

To properly support other types of pooling than discard pooling, we
need token spans, rather than just offsets. This change also adds for
each token its length in pieces, so that we have a full span
representation.