Simplify relation prediction in the biaffine parser
Before this change, we would use the following layers in the
biaffine parser:

1. Pairwise biaffine layer for predicting heads. Output:
   [batch_size, seq_len, seq_len]

2. Pairwise biaffine layer for predicting dependency labels. Output:
   [batch_size, seq_len, seq_len, n_rels]

We would first MST-decode the output of (1) to find the heads. Then we would
use the heads to gather the relevant relation predictions from (2). This
change replaces (2) by:

2b. Biaffine layer for predicting dependencies. Output:
    [batch_size, seq_len, n_rels].

To make (2b) effective, we give it as the input the dependent representations
of each tokens and the head representations gathered from the MST-decoded
output of (1).

The new approach is mathematically equivalent to the old one, but is simpler
and a bit faster.